net/udprelay{/endpoint}: move ServerEndpoint tests
Commit 0841477 moved ServerEndpoint to an independent package. Move its
tests over as well.

Updates tailscale/corp#27502
